pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-resolve"
version = "0.1.0"
description = "Provider-agnostic geocoding with persistent cache, rate limiting, and batch support"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.28",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
pandas = ["pandas>=2.0"]
all = ["pandas>=2.0"]
dev = ["pytest>=8.0"]

[project.scripts]
geo-resolve = "geo_resolve.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
